Restructure the integration tests with all queries and serializers in separate files
