+++ src/integration/mutations/AddConfusionTSMutation.graphql
@@ -1,0 +1,3 @@
+mutation AddConfusionTS($fp: ID, $sessionId: ID!, $difficulty: Int!, $speed: Int!) {
+  addConfusionTS(fp: $fp, sessionId: $sessionId, difficulty: $difficulty, speed: $speed)
+}
