+++ src/integration/mutations/AddFeedbackMutation.graphql
@@ -1,0 +1,3 @@
+mutation AddFeedback($fp: ID, $sessionId: ID!, $content: String!) {
+  addFeedback(fp: $fp, sessionId: $sessionId, content: $content)
+}
