+++ src/integration/mutations/AddResponseMutation.graphql
@@ -1,0 +1,3 @@
+mutation AddResponse($fp: ID, $instanceId: ID!, $response: QuestionInstance_ResponseInput!) {
+  addResponse(fp: $fp, instanceId: $instanceId, response: $response)
+}
