+++ src/integration/mutations/ArchiveQuestionsMutation.graphql
@@ -1,0 +1,6 @@
+mutation ArchiveQuestions($ids: [ID!]!) {
+  archiveQuestions(ids: $ids) {
+    id
+    isArchived
+  }
+}
