+++ src/integration/mutations/ArchiveQuestionsSerializer.js
@@ -1,0 +1,10 @@
+module.exports = {
+  test: ({ archiveQuestions }) => !!archiveQuestions,
+  print: ({ archiveQuestions }) => `
+    archiveQuestions: [${archiveQuestions.map(
+    ({ isArchived }) => `
+      isArchived: ${isArchived}
+    `,
+  )}]
+  `,
+}
