+++ src/integration/mutations/ChangePasswordMutation.graphql
@@ -1,0 +1,6 @@
+mutation ChangePassword($newPassword: String!) {
+  changePassword(newPassword: $newPassword) {
+    email
+    shortname
+  }
+}
