+++ src/integration/mutations/ChangePasswordSerializer.js
@@ -1,0 +1,9 @@
+module.exports = {
+  test: ({ changePassword }) => !!changePassword,
+  print: ({ changePassword: { email, shortname } }) => `
+      changePassword {
+        email: ${email}
+        shortname: ${shortname}
+      }
+    `,
+}
