+++ src/integration/mutations/CreateQuestionMutation.graphql
@@ -1,0 +1,54 @@
+mutation CreateQuestion(
+  $title: String!
+  $content: String!
+  $options: QuestionOptionsInput!
+  $solution: Question_SolutionInput
+  $type: Question_Type!
+  $tags: [ID!]!
+) {
+  createQuestion(
+    question: { title: $title, content: $content, options: $options, solution: $solution, type: $type, tags: $tags }
+  ) {
+    id
+    title
+    type
+    tags {
+      id
+      name
+    }
+    versions {
+      id
+      content
+      description
+      options {
+        SC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+        MC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+        FREE_RANGE {
+          restrictions {
+            min
+            max
+          }
+        }
+      }
+      solution {
+        SC
+        MC
+        FREE
+        FREE_RANGE
+      }
+      createdAt
+    }
+  }
+}
