+++ src/integration/mutations/CreateQuestionSerializer.js
@@ -1,0 +1,28 @@
+const { draftContentSerializer } = require('../../lib/test/serializers')
+
+module.exports = {
+  test: ({ createQuestion, modifyQuestion }) => !!createQuestion || !!modifyQuestion,
+  print: ({ createQuestion, modifyQuestion }) => {
+    const {
+      title, type, tags, versions,
+    } = createQuestion || modifyQuestion
+
+    return `
+    createQuestion / modifyQuestion {
+      title: ${title}
+      type: ${type}
+      tags: ${tags.map(tag => tag.name)}
+      versions: ${versions.map(
+    ({
+      content, description, options, solution,
+    }) => `
+        content: ${draftContentSerializer(content)}
+        description: ${description}
+        options: ${JSON.stringify(options)}
+        solution: ${JSON.stringify(solution)}
+      `,
+  )}
+    }
+  `
+  },
+}
