+++ src/integration/mutations/CreateSessionMutation.graphql
@@ -1,0 +1,31 @@
+mutation CreateSession($name: String!, $blocks: [Session_QuestionBlockInput!]!) {
+  createSession(session: { name: $name, blocks: $blocks }) {
+    id
+    confusionTS {
+      difficulty
+      speed
+    }
+    feedbacks {
+      id
+      content
+      votes
+    }
+    blocks {
+      id
+      status
+      instances {
+        id
+        question {
+          id
+          title
+          type
+        }
+      }
+    }
+    settings {
+      isConfusionBarometerActive
+      isFeedbackChannelActive
+      isFeedbackChannelPublic
+    }
+  }
+}
