+++ src/integration/mutations/DeleteFeedbackMutation.graphql
@@ -1,0 +1,10 @@
+mutation DeleteFeedback($sessionId: ID!, $feedbackId: ID!) {
+  deleteFeedback(sessionId: $sessionId, feedbackId: $feedbackId) {
+    id
+    feedbacks {
+      id
+      content
+      votes
+    }
+  }
+}
