+++ src/integration/mutations/LoginMutation.graphql
@@ -1,0 +1,3 @@
+mutation Login($email: String!, $password: String!) {
+    login(email: $email, password: $password)
+  }
