+++ src/integration/mutations/LogoutMutation.graphql
@@ -1,0 +1,3 @@
+mutation Logout {
+  logout
+}
