+++ src/integration/mutations/ModifyQuestionMutation.graphql
@@ -1,0 +1,55 @@
+mutation ModifyQuestion(
+  $id: ID!
+  $title: String
+  $content: String
+  $options: QuestionOptionsInput
+  $solution: Question_SolutionInput
+  $tags: [ID!]
+) {
+  modifyQuestion(
+    id: $id
+    question: { title: $title, content: $content, options: $options, solution: $solution, tags: $tags }
+  ) {
+    id
+    title
+    type
+    tags {
+      id
+      name
+    }
+    versions {
+      id
+      content
+      description
+      options {
+        SC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+        MC {
+          choices {
+            correct
+            name
+          }
+          randomized
+        }
+        FREE_RANGE {
+          restrictions {
+            min
+            max
+          }
+        }
+      }
+      solution {
+        SC
+        MC
+        FREE
+        FREE_RANGE
+      }
+      createdAt
+    }
+  }
+}
