+++ src/integration/mutations/PauseSessionMutation.graphql
@@ -1,0 +1,6 @@
+mutation PauseSession($id: ID!) {
+  pauseSession(id: $id) {
+    id
+    status
+  }
+}
