+++ src/integration/mutations/RegistrationMutation.graphql
@@ -1,0 +1,9 @@
+mutation CreateUser($email: String!, $password: String!, $shortname: String!, $institution: String!, $useCase: String) {
+  createUser(email: $email, password: $password, shortname: $shortname, institution: $institution, useCase: $useCase) {
+    id
+    email
+    shortname
+    institution
+    useCase
+  }
+}
