+++ src/integration/mutations/RegistrationSerializer.js
@@ -1,0 +1,15 @@
+module.exports = {
+  test: ({ createUser }) => !!createUser,
+  print: ({
+    createUser: {
+      email, shortname, institution, useCase,
+    },
+  }) => `
+    createUser {
+      email: ${email}
+      shortname: ${shortname}
+      institution: ${institution}
+      useCase: ${useCase}
+    }
+  `,
+}
