+++ src/integration/mutations/RequestPasswordMutation.graphql
@@ -1,0 +1,3 @@
+mutation RequestPassword($email: String!) {
+  requestPassword(email: $email)
+}
