+++ src/integration/mutations/UpdateSessionSettingsSerializer.js
@@ -1,0 +1,8 @@
+module.exports = {
+  test: ({ updateSessionSettings }) => !!updateSessionSettings,
+  print: ({ updateSessionSettings: { settings } }) => `
+    updateSessionSettings {
+      settings: ${JSON.stringify(settings)}
+    }
+  `,
+}
