+++ src/integration/queries/AccountSummaryQuery.graphql
@@ -1,0 +1,12 @@
+module.exports = `
+  query AccountSummary {
+    user {
+      id
+      shortname
+      runningSession {
+        id
+        runtime
+      }
+    }
+  }
+`
