+++ src/integration/queries/QuestionDetailsSerializer.js
@@ -1,0 +1,28 @@
+const { draftContentSerializer } = require('../../lib/test/serializers')
+
+module.exports = {
+  test: ({ question }) => !!question,
+  print: ({
+    question: {
+      title, type, instances, tags, versions,
+    },
+  }) => `
+    question {
+      title: ${title}
+      type: ${type}
+      tags: ${JSON.stringify(tags.map(tag => tag.name))}
+
+      instances: ${instances.length}
+      versions: ${versions.map(
+    ({
+      content, description, options, solution,
+    }) => `
+        content: ${draftContentSerializer(content)}
+        description: ${description}
+        options: ${JSON.stringify(options)}
+        solution: ${JSON.stringify(solution)}
+      `,
+  )}
+    }
+  `,
+}
