+++ src/integration/queries/RunningSessionQuery.graphql
@@ -1,0 +1,38 @@
+query RunningSession {
+  runningSession {
+    id
+    runtime
+    startedAt
+    activeStep
+    confusionTS {
+      difficulty
+      speed
+      createdAt
+    }
+    feedbacks {
+      id
+      content
+      votes
+      createdAt
+    }
+    blocks {
+      id
+      status
+      instances {
+        id
+        isOpen
+        version
+        question {
+          id
+          title
+          type
+        }
+      }
+    }
+    settings {
+      isConfusionBarometerActive
+      isFeedbackChannelActive
+      isFeedbackChannelPublic
+    }
+  }
+}
