+++ src/integration/queries/RunningSessionSerializer.js
@@ -1,0 +1,40 @@
+module.exports = {
+  test: ({ runningSession }) => !!runningSession,
+  print: ({
+    runningSession: {
+      activeStep, confusionTS, feedbacks, blocks, settings,
+    },
+  }) => `
+    runningSession {
+      activeStep: ${activeStep}
+      confusionTS: ${confusionTS.map(
+    ({ difficulty, speed }) => `
+        difficulty: ${difficulty}
+        speed: ${speed}
+      `,
+  )}
+      feedbacks: ${feedbacks.map(
+    ({ content, votes }) => `
+        content: ${content}
+        votes: ${votes}
+      `,
+  )}
+      blocks: ${blocks.map(
+    ({ status, instances }) => `
+        status: ${status}
+        instances: ${instances.map(
+    ({ isOpen, question, version }) => `
+          isOpen: ${isOpen}
+          version: ${version}
+          question {
+            title: ${question.title}
+            type: ${question.type}
+          }
+        `,
+  )}
+      `,
+  )}
+      settings: ${JSON.stringify(settings)}
+    }
+  `,
+}
